% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_to_beta.R
\name{beta_to_rate}
\alias{beta_to_rate}
\title{Convert a beta vector to a rate vector}
\usage{
beta_to_rate(beta)
}
\arguments{
\item{beta}{a beta vector.}
}
\value{
a rate vector.
}
\description{
Convert a beta vector to a rate vector.
}
\examples{
\donttest{
x <- beta_to_rate(c(0.7, 0.1, 0.2))
rate_to_beta(x)
}

}
