% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRRA.R
\name{CRRA}
\alias{CRRA}
\title{Constant Relative Risk Aversion (CRRA) Utility Function}
\usage{
CRRA(x, gamma, prob = rep(1/length(x), length(x)))
}
\arguments{
\item{x}{a payoff k-vector.}

\item{gamma}{the relative risk aversion coefficient.}

\item{prob}{a probability k-vector. By default, the states are assumed to
occur with equal probability.}
}
\value{
A list containing the following components:
\itemize{
\item u: the utility level.
\item CE: the certainty equivalent.
}
}
\description{
Compute the value and the certainty equivalent of the CRRA utility function.
}
\examples{
csv <- 0.05 # coefficient of standard deviation
mu <- 90 # mu <- 100
sigma <- mu * csv
x <- seq(mu - 5 * sigma, mu + 5 * sigma, length.out = 10000)
pd <- dnorm(x, mean = mu, sd = sigma)
gamma <- 0.8
# the ratio of risk premium to expected return (i.e. the relative risk premium).
(mu - CRRA(x, gamma, pd)$CE) / mu

####
df <- apply_expand.grid(
  function(arg) {
    CRRA(arg["x"], arg["gamma"])$u
  },
  x = seq(0.5, 3, 0.1),
  gamma = c(0.5, 1, 2, 3)
)
coplot(result ~ x | as.factor(gamma), data = df)
}
