% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_plot.R
\name{dst_plot}
\alias{dst_plot}
\title{Plot a Demand Structure Tree}
\usage{
dst_plot(dst)
}
\arguments{
\item{dst}{the demand structural tree.}
}
\description{
Plot a demand structure tree by the package DiagrammeR
and the type attribute of each node is shown.
}
