% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_tidy.R
\name{ge_tidy}
\alias{ge_tidy}
\title{Tidy a General Equilibrium}
\usage{
ge_tidy(ge, names.commodity, names.agent)
}
\arguments{
\item{ge}{the general equilibrium.}

\item{names.commodity}{a character vector consisting of
names of commodities.}

\item{names.agent}{a character vector consisting of
names of agents.}
}
\value{
A tidyed general equilibrium.
}
\description{
Add names to the matrices and vectors of a general equilibrium, and add
demand matrix, demand value matrix and supply value matrix to it.
}
