% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\title{Compute the Growth Rate}
\usage{
growth_rate(x, first.na = FALSE)
}
\arguments{
\item{x}{a vector or a matrix.}

\item{first.na}{If first.na==FALSE, the result doesn't contain the first NA.}
}
\value{
a vector or a matrix consisting of growth rates.
}
\description{
Compute the growth rates for a vector or each column of a matrix.
}
\examples{
\donttest{
x <- matrix(1:8, 4, 2)
growth_rate(x)
}
}
