% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePolicyTailAdjustment.R
\name{makePolicyTailAdjustment}
\alias{makePolicyTailAdjustment}
\title{Make a Policy of Tail Adjustment for a Timeline Model}
\usage{
makePolicyTailAdjustment(ind, gr = 0)
}
\arguments{
\item{ind}{a 2-column matrix or a numeric 2-vector that will be converted into a 2-column matrix.
In each row of the matrix, the first element corresponds to the index number of the last activity level of producing the product,
and the second element corresponds to the index number of a consumer who demands the product in the final period.}

\item{gr}{the growth rate.}
}
\value{
A policy, which is often used as an argument of the function sdm2.
}
\description{
Make a policy of tail adjustment for a timeline model.
Tail adjustment refers to the adjustment of the share coefficient of the last period of the consumer in the timeline model in order to let the model run in a steady-state equilibrium path.
}
\seealso{
{
\code{\link{makePolicyHeadTailAdjustment}};
\code{\link{makePolicyHeadAdjustment}}
}
}
