% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_plot.R
\name{node_plot}
\alias{node_plot}
\title{Plot a Tree and Show the Type Attribute}
\usage{
node_plot(node, ...)
}
\arguments{
\item{node}{a tree (i.e. a Node object).}

\item{...}{arguments to be to be passed to the function plot.Node.}
}
\description{
A wrapper of the function plot.Node of the packages data.tree.
If a non-leaf node has a type attribute, then the attribute will be shown.
}
\seealso{
\code{\link{demand_coefficient}}
}
