% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_print.R
\name{node_print}
\alias{node_print}
\title{Print a Tree and Its Fields}
\usage{
node_print(node, ...)
}
\arguments{
\item{node}{a Node object.}

\item{...}{arguments passed to print.Node.}
}
\description{
A wrapper of the function print.Node of the package data.tree.
Print a tree and its fields except the func field.
}
\examples{
\donttest{
dst <- node_set("firm", NA,
                type = "SCES",
                alpha = 2, beta = c(0.8, 0.2),
                es = 0.5,
                "wheat", "iron"
)

node_print(dst)

####
dst <- node_set("firm", NA,
                type = "FUNC",
                func = min,
                "wheat", "iron"
)

node_print(dst)
}

}
