% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_prune.R
\name{node_prune}
\alias{node_prune}
\title{Prune Nodes off a Tree by Names}
\usage{
node_prune(tree, ...)
}
\arguments{
\item{tree}{a tree (i.e. a Node object).}

\item{...}{some character strings specifies the names of nodes to be pruned.}
}
\value{
Invisibly returns the tree.
}
\description{
A wrapper of data.tree::Prunes. Prune nodes off a tree by names.
This function has side-effects, it modifies the tree given by the argument (see the package data.tree).
}
\examples{
\donttest{
dst <- node_set(
  "firm", NA,
  "lab", "cap", "land"
)
node_prune(dst, "cap", "land")
plot(dst)
}
}
