% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_to_beta.R
\name{rate_to_beta}
\alias{rate_to_beta}
\title{Convert a rate vector to a beta vector}
\usage{
rate_to_beta(rate)
}
\arguments{
\item{rate}{a rate vector.}
}
\value{
a beta vector.
}
\description{
Convert a rate vector to a beta vector.
}
\examples{
\donttest{
rate_to_beta(c(1, 1 / 3, 1 / 4))
}

}
