% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CES.R
\name{CES}
\alias{CES}
\title{CES Function}
\usage{
CES(sigma = 1 - 1/es, alpha, beta, x, theta = NULL, es = NA)
}
\arguments{
\item{sigma}{the sigma coefficient.}

\item{alpha}{the alpha coefficient.}

\item{beta}{a vector consisting of the beta coefficients.}

\item{x}{a vector consisting of the inputs.}

\item{theta}{a vector consisting of the theta coefficients.}

\item{es}{the elasticity of substitution. If es is not NA, the value of sigma will be ignored.}
}
\value{
The output or utility level.
}
\description{
CES function, e.g. alpha * (beta1 * (x1 / theta1)^sigma + beta2 * (x2 / theta2)^sigma)^(1 / sigma).
}
\examples{
\donttest{
CES(1, 1, c(0.4, 0.6), c(1, 1), c(0.4, 0.6))
}

}
